feat: Move wordlist
